retructure folders & move extensions back to diepAPI
