--- extensions/README.md
@@ -1,10 +1,0 @@
-# Extension
-
-Extensions are separate scripts that may be installed to extend diepAPIs functionality.
-
-For example if you want to use diepAPI.entityManager, install the entityManager extension.
-
-### Extensions:
-
-- Overlay
-- EntityManager
