--- extensions/overlay/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "module": "ESNext",
-        "moduleResolution": "node",
-        "baseUrl": "../../src",
-        "outDir": "../../dist"
-    },
-    "include": ["./", "../globals.d.ts"]
-}
