+++ src/apis/player_movement.ts
@@ -1,4 +1,4 @@
-import { Movement } from './core/movement';
+import { Movement } from '../core/movement';
 
 import { arena } from './arena';
 import { game } from './game';
