+++ src/extensions/README.md
@@ -1,0 +1,11 @@
+# Extension
+
+Extensions are modules that need to be loaded first to be useable (e.g. diepAPI.extensions.entityManager.load())
+
+
+### Extensions:
+
+- *entityManager
+
+
+*entityManager is an extension since it's use is performance heavy.
